Split Sources.py up into smaller files.
